requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
pydantic>=2.6.0
tqdm>=4.66.0
